+++ python/src/c2000_hil/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "VariableMatch",
 ]
 
-__version__ = "0.7.2"
+__version__ = "0.7.3"